+++ cyEngine/cyUtilities/source/cyVector4f.cpp
@@ -11,22 +11,22 @@
 namespace CYLLENE_SDK {
 
   Vector4f::Vector4f(const Vector2i& other)
-    : x(other.x),
-      y(other.y),
-      z(0),
-      w(0) {}
+    : x(static_cast<float>(other.x)),
+      y(static_cast<float>(other.y)),
+      z(0.0f),
+      w(0.0f) {}
 
   Vector4f::Vector4f(const Vector2f& other)
     : x(other.x),
       y(other.y),
-      z(0),
-      w(0) {}
+      z(0.0f),
+      w(0.0f) {}
 
   Vector4f::Vector4f(const Vector3f& other)
     : x(other[0]),
       y(other[1]),
       z(other[2]),
-      w(0) {}
+      w(0.0f) {}
   
   Vector4f::Vector4f(const Vector4f& other)
     : x(other.x),
